13feb

+++ src/main/java/es/seteruiz/tiernonews/controllers/ArticleController.java
@@ -3,6 +3,7 @@ package es.seteruiz.tiernonews.controllers;
 import es.seteruiz.tiernonews.models.Article;
 import es.seteruiz.tiernonews.services.ArticleService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -40,13 +41,27 @@ public class ArticleController {
     }
 
     @DeleteMapping("/article/{id}")
-    public void deleteArticleById(@PathVariable Long id){
-        articleService.deleteArticle(id);
+    public ResponseEntity deleteArticleById(@PathVariable Long id) {
+        if (articleService.existsById(id)) {
+            articleService.deleteArticle(id);
+            return ResponseEntity.ok("Deleted");
+        } else {
+            return ResponseEntity.status(400).body("No article found");
+        }
     }
 
-    @GetMapping("/article")
-    public void deleteArticleByTitle(@RequestParam String title){
-        articleService.deleteArticleByTitle(title);
+    //localhost:8080/article?title=tituloAEliminar
+    @DeleteMapping("/article")
+    public ResponseEntity deleteArticleByTitle(@RequestParam String title) {
+        int n = articleService.deleteArticleByTitle(title);
+        if (n > 0) {
+            //Retorno una respuesta http 200
+            return ResponseEntity.ok("Deleted " + n);
+        } else {
+            //Retorno una respuesta http 400
+            return ResponseEntity.status(400).body("No articles found");
+        }
     }
+
 
 }
